jwt 토큰에서 userId 받아오기, getBooks 수정

+++ controller/CartController.js
@@ -1,11 +1,13 @@
 import { StatusCodes } from "http-status-codes";
 import CartRepository from "../repository/CartRepository.js";
+import { getUserId } from "../middleware/jwt.js";
 
 const cartRepo = new CartRepository();
 
 export const getCart = async (req, res) => {
   try {
-    const { email, checked } = req.body;
+    const { checked } = req.body;
+    const email = getUserId(req.headers["authorization"]);
 
     const rows = await cartRepo.getCart(email, checked);
     return rows[0]
@@ -21,7 +23,8 @@ export const getCart = async (req, res) => {
 
 export const addToCart = async (req, res) => {
   try {
-    const { email, isbn, quantity } = req.body;
+    const { isbn, quantity } = req.body;
+    const email = getUserId(req.headers["authorization"]);
 
     const rows = await cartRepo.addToCart(email, isbn, quantity);
     const status =
@@ -36,8 +39,14 @@ export const addToCart = async (req, res) => {
 };
 
 export const deleteCartItem = async (req, res) => {
-  const { id } = req.params;
   try {
+    const email = getUserId(req.headers["authorization"]);
+    if (!email) {
+      return res
+        .status(StatusCodes.UNAUTHORIZED)
+        .json({ message: "로그인 필요" });
+    }
+    const { id } = req.params;
     const rows = await cartRepo.deleteCartItem(id);
     const statusCode =
       rows && rows.affectedRows ? StatusCodes.OK : StatusCodes.NOT_FOUND;
